remove compliation warnings

+++ src/value.cc
@@ -585,7 +585,7 @@ public:
   Value operator()(const std::string &str, const double &idx) const {
     int i = int(idx);
     Value v;
-    if (i >= 0 && i < str.size()) {
+    if ((i >= 0) && (i < (int)str.size())) {
       v = Value(str[int(idx)]);
       //      std::cout << "bracket_visitor: " <<  v << "\n";
     }
@@ -594,7 +594,7 @@ public:
 
   Value operator()(const Value::VectorType &vec, const double &idx) const {
     int i = int(idx);
-    if (i >= 0 && i < vec.size()) return vec[int(idx)];
+    if ((i >= 0) && (i < (int)vec.size())) return vec[int(idx)];
     return Value::undefined;
   }
